Add live reloading for dev
This commit adds bind mounts for the ui and api folders so that changes to the contents affect the containers.
Additionally, both the ui and api have been set to hot reload:
The ui is using the webpack dev server/hmr setup from CRA
The api is using the new `--watch` option added in Node 18 to do hot reloading.