Fix Apollo client error in App.test.js
This commit adds Apollos MockedProvider component to the App test to
resolve the error that was being produced.
Components using Apollo client depend on the client to be available via
 the context.
This client is used to fetch data from the API.
When testing the MockedProvider is wrapped around components that
need a client and test data is passed to the mocks prop allowing
the MockedProvider to simulate an interaction with the API.
As part of this refactor the queries are extracted into a separate
file to allow imports into both the tests and the GetAll component.